uan: (fixes #2386) Move UanPhyGen out of Tx mode when energy depleted
This patch adds a new PHY state 'DISABLED' and moves the Phy to this
state when energy is depleted, cancelling any pending transmissions or
receptions, and preventing further energy consumption below the low battery
threshold from the energy source.  It also introduces new methods to handle
the event when energy is recharged (to move the state out of DISABLED) by
adding an EnergyRechargeHandler callback.  Federico Guerra also contributed
to the development of this patch.