test:  Reduce scope of valgrind example tests to speed execution time
Some example programs are tested under different configurations; as long
as we test one of these configurations, there is little to gain by valgrind
testing additional configurations, but './test.py -g' will take longer.  So
in general, only one or a limited number of configurations of an example
are valgrind tested.